Added generation of random powder data.
A number of random Bragg peaks are defined with a given range of
d spacings, delta d over d values and relative strengths.
The wavelength frame is defined and the Bragg peaks available for
each pixel is identified. A poisson distributed random number of
events are then drawn from each bragg peak, each with a gaussion
distribution of d spacings. The instrument length, detector
distance and a random source origin time is used to calculate the
corresponding time for the event.

This provides a dataset with somewhat realistic correlations
between time and theta value of each pixel. There is no background,
but this could be added if desired.

The performance scaling of the code is problematic for the number
of pixels generated, but the scaling for number of events is good.

1E7 events in 1E5 pixels takes 15 seconds on my laptop. Going to
5E5 pixels the run time is around 6 minutes.