Resulting code from DAS-148 where the Mantid function SmoothData was ported to scipp.
SmoothData updates each point to an average of the surrounding points and itself.

The Mantid implementation uses a rolling average, where floating point errors can become problematic, especailly for the errors that ues a squareroot which has to use absolute values to avoid imaginary errors.
Meanwhile the scipp version takes the mean of each point set, which is slower but both simpler and has no issues with accumulation of floating point error.

A Jupyter Notebook demo is available that shows the implementations produce eqvivalent results.