Update imaging bench to do direct comparisons in same notebook
Updates the imaging bench to do direct comparisons between Mantid and
Scipp in the same notebook. This also highlights where we have feature
parity, and the complexity of each to do certain steps.

An option is included to select which implementation we test (or both),
as ~64GB is required to bench both currently.