q resolution steps
Basics implemenation based on math present in mantid TOFSANSResolutionByPixel. This is not yet producing the same results as Mantid. That will need to be invesigated next. Also the implementation presented here is very slow with no attention whatsoever applied to performance.

There are some additional units for scipp that need to be supported for this to work. namely inverse velociy units s/m.