Use SciPy for spectra mask and median filters
Switch to using a convolutional filter to check for adj. spectra points
so we can determine where we should replace with a True / False value
without having to loop over the entire dataset

Replaces the custom median filter with a drop in median filter provided
in SciPy. The old one is currently included whilst data validation is
completed