Added validators
This updates ProcessWidget so that the conversion functions
provided for input cells can optionally do a validation check
and throw a ValueError if this fails. This will then be handled
as a user input error by the widget. An alternative route to
pass in optional validators would be to have a seperate dict
sharing keys with the input dict which specifies options to use.
I went for this method over that because conversion and validation are often
closely linked.