Added autocomplete options.
Added ability to optionally specify autocomplete values for cells.
This is demonstrated for the data inputs and dimension values in the
conversion widgets. Note however that without linking up some observer
you have to re-run the cell to get the latest data to properly appear as
options. In the long term if may be worth having a notifier control object
which all widgets which create scipp objects register with and can then subscribe
to if they wish to know when scipp objects are created by have not done so currently.